% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cor.R
\name{check_cor}
\alias{check_cor}
\title{Check pairwise correlations in a data frame}
\usage{
check_cor(x)
}
\arguments{
\item{x}{A data frame containing variables for which to compute correlations.}
}
\value{
A message indicating whether all pairwise correlations are positive.
}
\description{
This function checks whether all pairwise correlations between variables in a
data frame are positive.
}
