% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invalid_key.R
\name{check_invalid_key}
\alias{check_invalid_key}
\title{Check if a vector contains valid keys for specifying invalid response group factor loadings}
\usage{
check_invalid_key(x)
}
\arguments{
\item{x}{A numeric vector to be checked.}
}
\value{
Returns \code{NULL} if the vector contains both positive and negative
values. Otherwise, raises an error.
}
\description{
This function checks if a numeric vector contains valid keys for specifying
factor loadings for an invalid response group in a factor mixture model.
Specifically, the vector must contain both positive and negative values.
}
