% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_observed.R
\name{check_observed}
\alias{check_observed}
\title{Check if a vector of observed variables is valid}
\usage{
check_observed(x)
}
\arguments{
\item{x}{A vector of observed variables to check.}
}
\value{
Nothing; this function is used for its error-checking side effects.
}
\description{
This function checks whether a vector of observed variables is valid for use
detecting carelessness. Specifically, it checks whether the input is a
character vector with at least two elements.
}
