% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{construct_model}
\alias{construct_model}
\title{Construct Factor Mixture Model}
\usage{
construct_model(data, ov, reversed)
}
\arguments{
\item{data}{A data frame or matrix containing observed variable data.}

\item{ov}{A character vector of the observed variable names.}

\item{reversed}{A character vector of reverse coded observed variables that
have already been reverse-coded and will have their loadings reversed in
the invalid class (variables must be a subset of ov).}
}
\value{
Returns an OpenMx factor mixture model object.
}
\description{
This function constructs an OpenMx factor mixture model to identify careless
responses.Note that function defaults to assuming prior reverse-coding of
reversed items (i.e., that all items will have positive loadings in the valid
response group).
}
