% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_assignment_probabilities.R
\name{get_class_assignment_probabilities}
\alias{get_class_assignment_probabilities}
\title{Get class assignment probabilities from factor mixture model}
\usage{
get_class_assignment_probabilities(fit)
}
\arguments{
\item{fit}{A factor mixture model fit object obtained from
\code{estimate_model}.}
}
\value{
A data frame with two columns, "valid" and "invalid", containing the
estimated probabilities of each observation belonging to the "valid" and "invalid"
latent classes, respectively.
}
\description{
Given a factor mixture model fit object, returns a data frame containing the
estimated probabilities of each observation belonging to the valid and
invalid response classes.
}
