% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assignment_probability_distribution.R
\name{plot_assignment_probability_distribution}
\alias{plot_assignment_probability_distribution}
\title{Plot assignment probability distribution}
\usage{
plot_assignment_probability_distribution(probs, cutoff = 0.95)
}
\arguments{
\item{probs}{A data frame containing classification probabilities. The data
frame must have two columns, one containing probabilities of the valid
response class and one containing probabilities of the invalid response
class.}

\item{cutoff}{An optional scalar indicating the probability cutoff above
which to calculate the proportion of responses. Default value is 0.9.}
}
\value{
A histogram of the probability of invalid response class assignment
  and a plot of the number of responses with probabilities exceeding the
  given cutoff (default = 0.9).
}
\description{
Given a data frame of classification probabilities, this function produces a
histogram of the probability of invalid responses and displays the number of
responses with probabilities exceeding a given cutoff.
}
\examples{
probs = data.frame(valid = runif(100))
probs$invalid = 1-probs$valid
plot_assignment_probability_distribution(probs)
}
