% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decision_threshold.R
\name{plot_decision_threshold}
\alias{plot_decision_threshold}
\title{Plot decision threshold}
\usage{
plot_decision_threshold(probs)
}
\arguments{
\item{probs}{A data frame containing classification probabilities. The data
frame must have two columns, one containing probabilities of the valid
class and one containing probabilities of the invalid class.}
}
\value{
A plot of decision thresholds.
}
\description{
Given a data frame of classification probabilities, this function produces a
plot of decision thresholds.
}
\examples{
probs <- data.frame(valid = runif(100), invalid = runif(100))
plot_decision_threshold(probs)
}
