% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_classes.R
\name{specify_classes}
\alias{specify_classes}
\title{Define the valid and invalid response classes in the carelessness factor
mixture model.}
\usage{
specify_classes(
  ov,
  valid_loadings,
  valid_latent_variance,
  invalid_loadings,
  invalid_latent_variance,
  latent_means,
  error_var,
  item_int,
  fit_fun = NULL,
  lv = "G"
)
}
\arguments{
\item{ov}{A character vector of observed variable names.}

\item{valid_loadings}{An OpenMx path object specifying the valid group's
factor loadings.}

\item{valid_latent_variance}{An OpenMx path object specifying the valid
group's latent variable variance.}

\item{invalid_loadings}{An OpenMx path object specifying the invalid group's
factor loadings.}

\item{invalid_latent_variance}{An OpenMx path object specifying the invalid
group's latent variable variance.}

\item{latent_means}{An OpenMx path object specifying the latent variable
means.}

\item{error_var}{An OpenMx path object specifying the error variances.}

\item{item_int}{An OpenMx path object specifying the item intercepts.}

\item{fit_fun}{An OpenMx fit function. Default is NULL, which specifies
maximum likelihood estimation.}

\item{lv}{A character string indicating the name of the latent variable.
Default is "G".}
}
\value{
A list containing the two class definitions.
}
\description{
Define the valid and invalid response classes in the carelessness factor
mixture model.
}
