% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_invalid_loadings.R
\name{specify_invalid_loadings}
\alias{specify_invalid_loadings}
\title{Specify factor loadings for invalid response group in factor mixture model}
\usage{
specify_invalid_loadings(ov, invalid_key, lv = "G", ...)
}
\arguments{
\item{ov}{A character vector of observed variable names for the invalid
response group.}

\item{invalid_key}{A vector of numeric values specifying the factor loadings
reverse-coded items in the invalid response group. The vector must contain
positive and negative values (e.g., \code{c(1, -1, 1, -1)}).}

\item{lv}{The name of the latent variable for the invalid response group.
Defaults to "G".}

\item{...}{Additional arguments passed to \code{\link{mxPath}}.}
}
\value{
A path (factor loading) object in OpenMx.
}
\description{
This function creates an OpenMx path object, specifying factor loadings for
the invalid response group in a factor mixture model. The factor loadings are
fixed to the values specified by the \code{invalid_key} argument.
}
\examples{
# Create a factor with loadings of c(1, -1, 1, -1) for "y4", "y5", "y6", and "y7"
specify_invalid_loadings(c("y4", "y5", "y6", "y7"), invalid_key = c(1, -1, 1, -1))

}
