% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_latent_mean.R
\name{specify_latent_mean}
\alias{specify_latent_mean}
\title{Specify the mean of a latent variable in factor mixture modeling}
\usage{
specify_latent_mean(x = 0, lv = "G", ...)
}
\arguments{
\item{x}{A numeric value specifying the mean of the latent variable. Defaults
to 0.}

\item{lv}{The name of the latent variable. Defaults to "G".}

\item{...}{Additional arguments passed to \code{\link{mxPath}}.}
}
\value{
A path object in OpenMx.
}
\description{
This function creates a path object in OpenMx, specifying the mean of a latent
variable in a factor mixture model. The mean is fixed to the value specified
by the \code{x} argument.
}
\examples{
# Create a factor with a mean of 1 for "G"
specify_latent_mean(x = 1, lv = "G")

}
