% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_latent_variance.R
\name{specify_latent_variance}
\alias{specify_latent_variance}
\title{Specify the variance of a latent variable in a factor mixture model}
\usage{
specify_latent_variance(x = 1, lv = "G")
}
\arguments{
\item{x}{A numeric value specifying the variance of the latent variable.
Defaults to 1.}

\item{lv}{The name of the latent variable for which to specify the variance.
Defaults to "G".}
}
\value{
A path (variance) object in OpenMx.
}
\description{
This function creates an OpenMx path object, specifying the variance of a
latent variable in a factor mixture model. The variance is fixed to the value
specified by the \code{x} argument.
}
\examples{
# Create a variance of 0.5 for the latent variable "G"
specify_latent_variance(x = 0.5, lv = "G")

}
