% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_loadings.R
\name{specify_loadings}
\alias{specify_loadings}
\title{Specify factor loadings in factor mixture modeling}
\usage{
specify_loadings(ov, key, lv, ...)
}
\arguments{
\item{ov}{A character vector of observed variable names.}

\item{key}{A vector of numeric values specifying the factor loadings for
the observed variables in \code{ov}.}

\item{lv}{The name of the latent variable for the group of observed variables.
Defaults to "G".}

\item{...}{Additional arguments passed to \code{\link{mxPath}}.}
}
\value{
A path (factor loading) object in OpenMx.
}
\description{
This function creates a path object in OpenMx, specifying factor loadings for
a set of observed variables in a factor mixture model. The factor loadings
are fixed to the values specified by the \code{key} argument.
}
