% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_observed_intercepts.R
\name{specify_observed_intercepts}
\alias{specify_observed_intercepts}
\title{Specify intercpts for observed variables in factor mixture modeling}
\usage{
specify_observed_intercepts(ov, ...)
}
\arguments{
\item{ov}{A character vector of observed variable names.}

\item{...}{Additional arguments passed to \code{\link{mxPath}}.}
}
\value{
A path (variance) object in OpenMx.
}
\description{
This function creates a path object in OpenMx, specifying freely estimated
intercpts for observed variables in a factor mixture model.
}
\examples{
# Create a residual variance of 1 for "y1", "y2", and "y3"
specify_observed_intercepts(c("y1", "y2", "y3"))

}
